Trivia

+++ generic/com_vanilla.c
@@ -157,8 +157,8 @@ initialize_machine(int *argc, char ***argv)
   }
 #else
   if(sizeof(unsigned int)!=4) {
-    printf("node %i: SHORT_IS_32BIT is not set but sizeof(unsigned int)=%i\n",
-	   mynode(), sizeof(unsigned int));
+    printf("node %i: SHORT_IS_32BIT is not set but sizeof(unsigned int)=%d\n",
+	   mynode(), (int)sizeof(unsigned int));
     terminate(1);
   }
 #endif
